Move jsdom test fixtures to core package
Several other tests require a virtual dom otherwise browser exclusive
errors are reported. For instance keybinding.spec.ts needs to have
KeyboardEvent registered, and upcoming frontend preferences tests need
to have the navigator defined (because of certain imports).

It also makes it much easier to develop mocha tests without using
absolute imports as before, if you imported something from an index
file, and this index file would in turn export something else like
`application-shell.ts`, which also includes some phosphor libraries,
then you would have errors about undefined navigator as phosphor needs
those. That means you could have those errors even though you never
needed phosphor directly in your tests. The workaround was to directly
specify the path of the dependency, but that isn't ideal as you lose the
advantage of using index.ts files.
