+++ packages/preview/package.json
@@ -50,9 +50,7 @@
   "devDependencies": {
     "@theia/ext-scripts": "^0.2.0",
     "@types/chai-string": "^1.1.31",
-    "@types/jsdom": "^11.0.4",
-    "chai-string": "^1.4.0",
-    "jsdom": "^11.5.1"
+    "chai-string": "^1.4.0"
   },
   "nyc": {
     "extends": "../../configs/nyc.json"
