+++ packages/preview/src/browser/markdown/markdown-preview-handler.spec.ts
@@ -5,7 +5,7 @@
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-import { cleanupJSDOM } from '../../common/test/jsdom';
+import { cleanupJSDOM } from '@theia/core/lib/browser/test/jsdom';
 
 import * as chai from 'chai';
 import { expect } from 'chai';
